Fixed bug
When the button has been clicked, it has focus in the browser. The button would be the selected element.
Then, if the enter key is pressed, this triggers the default action on the button, which is a click event.

To prevent the click event being triggered = event.preventDefault().
Very interesting